+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 - new properties for `openpile.construct.Pile`: `weight` and `volume`
 - new `openpile.construct.Pile` method: `set_I()` to change the second moment of area of a given pile segment
 - new `SoilProfile.plot()` method to visualize the soil profile
-- 
+- API sand and API clay curves and models now accept `kind` instead of `Neq` arguments to differentiate between 
+  static and cyclic curves
 
 ## [0.1.0] - 2023-04-10
 ### Added